While the recursive search is active, it is no longer possible to select files; The correct number of files found is now displayed in the footer depending on the search type

+++ frontend/src/components/IFMTable.vue
@@ -131,6 +131,8 @@ async function handleFileNavigation(file) {
 
 /* Allows the individual selection of multiple files */
 function toggleFileSelection(file, event) {
+    if (filesStore.isRecursiveSearch) return;
+
     if (event.ctrlKey || event.type === 'dblclick') {
         if (filesStore.selectedFiles.has(file.name)) {
             filesStore.selectedFiles.delete(file.name);
